move damsel code into its own method, implement damsel spawn fix
